+++ src/components/LoanFinding/PartnerSpotlightSection.vue
@@ -1,35 +1,41 @@
 <template>
-	<div class="tw-w-full tw-bg-secondary">
+	<div v-if="mfiRecommendationsLoans.length > 0" class="tw-w-full tw-bg-secondary">
 		<div class="tw-mx-auto tw-px-0 md:tw-px-4 lg:tw-px-8" style="max-width: 1200px;">
 			<m-f-i-hero />
+		</div>
 
-			<mfi-loans-wrapper
-				v-if="selectedChannelLoanIds.length > 0"
-				:selected-channel-loan-ids="selectedChannelLoanIds"
-				:selected-channel="selectedChannel"
-				:show-tags="true"
-				class="tw-py-4"
-			/>
+		<lending-category-section
+			title="Support Borrowers from Fundación Pro Mujer"
+			:loans="mfiRecommendationsLoans"
+		/>
+
+		<div class="tw-flex tw-justify-center tw-mt-4">
+			<a
+				class="tw-text-action tw-text-base"
+				href="https://www.kiva.org/lend/filter?partner=59"
+				v-kv-track-event="['event-tracking', 'click', 'mfi-view-all']"
+			>
+				View All
+			</a>
 		</div>
 	</div>
 </template>
 
 <script>
 import MFIHero from '@/components/LoansByCategory/MFIRecommendations/MFIHero';
-import MfiLoansWrapper from '@/components/LoansByCategory/MFIRecommendations/MfiLoansWrapper';
+import LendingCategorySection from '@/components/LoanFinding/LendingCategorySection';
 import mfiRecommendationsLoans from '@/graphql/query/lendByCategory/mfiRecommendationsLoans.graphql';
 
 export default {
 	name: 'PartnerSpotlightSection',
 	components: {
 		MFIHero,
-		MfiLoansWrapper
+		LendingCategorySection
 	},
 	inject: ['apollo', 'cookieStore'],
 	data() {
 		return {
 			mfiRecommendationsLoans: [],
-			selectedChannel: {},
 		};
 	},
 	computed: {